feat : 이미지 전송 로딩 화면

+++ wHereRU_Spring/src/main/java/kosa/hdit5spring/whereru/notice/controller/NoticeController.java
@@ -36,7 +36,7 @@ public class NoticeController {
 	}
 	
 	@PostMapping("setLocation")
-	public void setLocation(@RequestBody List<float> locationList) {
+	public void setLocation(@RequestBody List<Double> locationList) {
 		System.out.println("------------------------------------------");
 		System.out.println("위도 :" +locationList.get(0)+"경도 :"+locationList.get(1));
 		System.out.println("------------------------------------------");
